rename ResponsePage.vue
