+++ src/components/ProblemDetailsPage.vue
